feat: se agrega gestión de productos

+++ backend/src/controllers/cloudinary.controller.js
@@ -1,12 +1,19 @@
 const cloudinary = require('cloudinary').v2;
 const fs = require('fs');
+const path = require('path');
 
 const uploadImage = async (req, res) => {
   try {
     if (!req.file) {
       return res.status(400).json({ error: 'No se envió ninguna imagen' });
     }
-    if (!req.file.mimetype.startsWith('image/')) {
+    if (
+      !req.file.mimetype.startsWith('image/') &&
+      !(
+        req.file.mimetype === 'application/octet-stream' &&
+        ['.jpg', '.jpeg', '.png', '.gif', '.webp'].includes(path.extname(req.file.originalname).toLowerCase())
+      )
+    ) {
       fs.unlinkSync(req.file.path);
       return res.status(400).json({ error: 'Solo se permiten imágenes' });
     }
@@ -45,9 +52,19 @@ const getImagesFromFolder = async (req, res) => {
   }
 };
 
-
+const deleteImage = async (req, res) => {
+  try {
+    const { public_id } = req.body;
+    if (!public_id) return res.status(400).json({ error: 'public_id requerido' });
+    await cloudinary.uploader.destroy(public_id);
+    res.json({ message: 'Imagen eliminada' });
+  } catch (error) {
+    res.status(500).json({ error: 'Error al eliminar la imagen de Cloudinary' });
+  }
+};
 
 module.exports = {
   getImagesFromFolder,
-  uploadImage
+  uploadImage,
+  deleteImage
 };
